xfpga: use single source of truth for mmap flags
Previously, the mmap flags used by the huge page allocation code
we replicated in several of the source files and even in some of
the test code. Move the flags to common_int.h and use that as the
single definition of the flags.

Allocation requests for 2MB hugepages were being satisfied from
the kernel's default huge page pool. Change the 2MB allocation flags
to explicitly request these of the 2MB pool.