xfpga: fix irq use in events API
The events API was always using index 0 when registering for interrupts.
Use the number of interrupts available along with a bitset of interrupts
already set for validating registering events. Also, save the interrupt
index in the event_handle data structure for associating interrupt
number with the event. This is used when unregistering for events.