pybind11: add el7 guard in selecting pybind11 tag
pybind11 latest tag no longer builds on default el7 machines, at it has bumped
up the minimum supported cmake version.

https://github.com/pybind/pybind11/releases/tag/v2.6.0

This means that el7 won't compile a pybind11 version which supports python 3.9
(this is rationale for bumpding pybind11 version in the first place). Howver, that
is likely out of scope of el7 - that support is on upstream Fedora systems for
example.

Signed-off-by: Marroquin, Asgard <asgard.marroquin@intel.com>