cmake: packaging: add opae_python_install helper function
Prepare to unify installation of python modules/utils by creating a function
to be used by such.

The function takes care of installation path, so that `make
DESTDIR=/path/to/root` and `cmake .. -DCMAKE_INSTALL_PREFIX` are
respected. It also supports RPM packaging which can use the generated list
of installed files.

Signed-off-by: Martin Hundebøll <mhu@silicom.dk>