vfio-plugin: fix get_guid to use 64bit rd
Current implementation of get_guid copies the input bytes (uint64_t *h)
in reverse order but because it does so per byte, the result is 32-bit
MMIO accesses. This change dereferences the two 64-bit words and calls
bswap_64 to have the bytes swapped and assigns the result to the output
guid variable.

Signed-off-by: Rodrigo Rojo <rodrigo.rojo@intel.com>